DOPE-234: renaming
