+++ crystal-map-connector/src/main/kotlin/ch/ergon/dope/extension/clause/joinhint/KeysHint.kt
@@ -1,10 +1,10 @@
-package ch.ergon.dope.extension.clause
+package ch.ergon.dope.extension.clause.joinhint
 
 import ch.ergon.dope.resolvable.clause.model.joinHint.keysHint
 import ch.ergon.dope.toDopeType
 import com.schwarz.crystalapi.schema.CMJsonField
 import com.schwarz.crystalapi.schema.CMJsonList
 
-fun keysHint(useKeys: CMJsonField<String>) = keysHint(useKeys.toDopeType())
+fun keysHint(keys: CMJsonField<String>) = keysHint(keys.toDopeType())
 
-fun keysHint(useKeys: CMJsonList<String>) = keysHint(useKeys.toDopeType())
+fun keysHint(keys: CMJsonList<String>) = keysHint(keys.toDopeType())
