+++ crystal-map-connector/src/test/kotlin/ch/ergon/dope/extensions/clause/SelectClauseTest.kt
@@ -4,7 +4,7 @@ import ch.ergon.dope.DopeQueryManager
 import ch.ergon.dope.extension.clause.groupBy
 import ch.ergon.dope.extension.clause.innerJoin
 import ch.ergon.dope.extension.clause.join
-import ch.ergon.dope.extension.clause.keysHint
+import ch.ergon.dope.extension.clause.joinhint.keysHint
 import ch.ergon.dope.extension.clause.leftJoin
 import ch.ergon.dope.extension.clause.limit
 import ch.ergon.dope.extension.clause.offset
